Refactored the Samples. Signed-off-by: Dileep Miriyala <m.dileep@gmail.com>
