--- NFugue.Samples/Program.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NFugue.Samples
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            //TO DO: To be improved.
-            HelloWorld.Go();
-            HelloWorld2.Go();
-            IntroToPatterns.Go();
-            IntroToPatterns2.Go();
-            IntroToChordProgressions.Go();
-            AdvancedChordProgressions.Go();
-            TwelveBarBlues.Go();
-            IntroToRhythms.Go();
-            AdvancedRhythms.Go();
-            CarnaticReplacementMapDemo.Go();
-            SolfegeReplacementMapDemo.Go();
-            LSystemMusic.Go();
-        }
-    }
-}
