+++ src/NFugue.Samples/App.config
