+++ src/NFugue.Samples/MyParserListener.cs
@@ -1,0 +1,19 @@
+﻿namespace NFugue.Samples
+{
+
+    class MyParserListener //: ParserListenerAdapter
+    {
+
+        //public int counter;
+
+
+        //public override void onNoteParsed(Note note)
+        //{
+        //    // A "C" note is in the 0th position of an octave
+        //    if (note.GetPositionInOctave() == 0)
+        //    {
+        //        counter++;
+        //    }
+        //}
+    }
+}
