+++ src/NFugue.Samples/ParserDemo2.cs
@@ -1,0 +1,29 @@
+﻿namespace NFugue.Samples
+{
+
+
+    /*	Create a Listener to Find Out About Music
+    You can create a ParserListener to listen for any musical event
+    that any parser is parsing. Here, we'll create a simple tool that
+    counts how many "C" notes (of any octave) are played in any song.
+    */
+
+    public class ParserDemo2
+    {
+
+        public static void Run()
+
+        {
+
+
+            //String fileName = "PUT A MIDI FILE HERE";
+            //MidiParser parser = new MidiParser(); // Remember, you can use any Parser!
+            //MyParserListener listener = new MyParserListener();
+
+            //parser.AddParserListener(listener);
+            //parser.Parse(MidiSystem.getSequence(new File(fileName)));
+
+            //Console.WriteLine("There are " + listener.counter + " 'C' notes in this music.");
+        }
+    }
+}
