+++ src/NFugue.Samples/Program.cs
@@ -1,0 +1,11 @@
+﻿namespace NFugue.Samples
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            new SampleRunner().Run();
+
+        }
+    }
+}
