+++ src/NFugue.Samples/Properties/AssemblyInfo.cs
