+++ src/NFugue.Samples/packages.config
